+++ src/extension.ts
@@ -10,11 +10,26 @@ let compilerWatch;
 
 // this method is called when your extension is activated
 export function activate(context: vscode.ExtensionContext) {
-    if (!config.enabled) {
-        console.log('vscode-mocha-test-runner is not enabled - enable it by "mocha.enabled" = true in settings.');
-        return;
+    if (config.enabled === true) {
+        doActivate(context);
+    } else if (config.enabled !== false) {
+        vscode.window.showInformationMessage('vscode-mocha-test-runner is not enabled', 'Enable', 'Do not show again')
+            .then(result => {
+                if (result === 'Enable') {
+                    config.update('enabled', true);
+                    doActivate(context);
+                } else if (result === 'Do not show again') {
+                    config.update('enabled', false);
+                }
+            });
     }
+}
 
+// this method is called when your extension is deactivated
+export function deactivate() {
+}
+
+function doActivate(context: vscode.ExtensionContext) {
     try {
         codeLensProvider = new TestsCodeLensProvider();
         outputChannel = vscode.window.createOutputChannel('Mocha test runner');
@@ -31,8 +46,4 @@ export function activate(context: vscode.ExtensionContext) {
     } catch (err) {
         console.error('vscode-mocha-test-runner activate error: ' + err);
     }
-}
-
-// this method is called when your extension is deactivated
-export function deactivate() {
 }